+++ src/com/intellij/idea/plugin/hybris/project/compile/ProjectBeforeCompilerTask.kt
@@ -50,6 +50,7 @@ import kotlin.io.path.exists
 import kotlin.io.path.extension
 import kotlin.io.path.name
 
+// TODO: add progress indicator
 class ProjectBeforeCompilerTask : CompileTask {
 
     override fun execute(context: CompileContext): Boolean {